msm: eva: Avoid mutex lock while holding spin lock
spin_lock disables preemption. If we try to acquire the mutex while
holding the spin lock and if it gets blocked due to the mutex already being
acquired by another thread, then this will lead the current thread to go
into blocked state but CPU can't schedule another thread as preemption
is disabled by spin lock. This is a deadlock now.
Mutex lock is now not allowed inside the spin lock.

Change-Id: I9c9c4a4fda2564890743381af923614b54158a55
Signed-off-by: Aniruddh Sharma <quic_anirshar@quicinc.com>