msm: eva: Get bus info based on DT node label
Bus info was being obtained based on the position of the cvp-ddr
bus node in the .dtsi file, which cannot be relied on as the order
of bus nodes can change in the compiled .dtb file.
This led to incorrect or abnormally high BW votes.
Ported fixes to get info based on the bus node's label instead of
the node's position from Aurora LA.

Signed-off-by: Shivani Hanji <quic_shanji@quicinc.com>

Change-Id: I79257e3f7a9125f00550e9d7ad6e2396c76bded5